Add a fast lookup mechanism for known ICAO addresses.
This was an obvious thing to do, but for some reason I didn't think of doing it.
There is now both a list and a bitmap to contain the list of known aircraft. The
latter can be queried very quickly. This speeds up CRC checking and makes error
correction feasible for messages using XORed CRCs.

This commit also adds double bit error correction for messages not using XORed
CRCs.

Also some whitespace removal to keep git happy.