Add the beginnings of the aircraft database code.
The eventual aim is to have a database of aircraft objects which contain
all the known information about each aircraft. The objects will be passed
mode S / ADS-B messages and will update their status accordingly.

This commit adds aircraft.py which contains a basic template of a simple
aircraft object and db_adsb.py. The latter is like dump_adsb.py except
that it builds a dictionary of aircraft objects and dumps the accumulated
data from each one at EOF. Both of these files are works in progress.